Unifies event naming and splits the OnLoadEvent to BeforeLoadEvent and AfterLoadEvent
